move ProxyClass into separate file
